+++ includes/system/class-ip.php
@@ -81,6 +81,7 @@ class IP {
 	 * @since 1.0.0
 	 */
 	public static function get_current() {
+		$ip = '127.0.0.1';
 		for ( $i = 0 ; $i < 2 ; $i++ ) {
 			foreach (
 				[
@@ -93,9 +94,6 @@ class IP {
 			) {
 				if ( array_key_exists( $field, $_SERVER ) ) {
 					$ip = self::maybe_extract_ip( explode( ',', filter_input( INPUT_SERVER, $field ) ), 1 === $i );
-					if ( '' !== $ip ) {
-						return $ip;
-					}
 				}
 			}
 			foreach (
@@ -109,13 +107,10 @@ class IP {
 			) {
 				if ( array_key_exists( $field, $_SERVER ) ) {
 					$ip = self::maybe_extract_ip( array_reverse( explode( ',', filter_input( INPUT_SERVER, $field ) ) ), 1 === $i );
-					if ( '' !== $ip ) {
-						return $ip;
-					}
 				}
 			}
 		}
-		return '127.0.0.1';
+		return $ip;
 	}
 
 	/**