+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to **DecaLog** are documented in this *changelog*.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and **DecaLog** adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased - will be 2.2.0]
+## [2.2.0] - 2020-10-11
 
 ### Added
 - DecaLog now integrates [Spyc](https://github.com/mustangostang/spyc) as yaml parser.